requirements.txt
annotated-types==0.6.0
black==23.12.1
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
defusedxml==0.7.1
emoji==2.8.0
fonttools==4.46.0
fpdf2==2.7.7
idna==3.4
inflect==7.0.0
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
Pillow==10.1.0
platformdirs==4.1.0
pluggy==1.3.0
pydantic==2.4.2
pydantic_core==2.10.1
pyfiglet==0.7
pytest==7.4.4
requests==2.31.0
tabulate==0.9.0
typing_extensions==4.8.0
urllib3==2.0.7
